// src/app/api/model/episodeDto.ts
/**
 * Rezflix Backend
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EpisodeStatusDto } from './episodeStatusDto';


export interface EpisodeDto { 
    watchOrder: number;
    title?: string;
    id: string;
    status: EpisodeStatusDto;
    hlsLink?: string;
}