// src/app/api/model/episodeStatusDto.ts
/**
 * Rezflix Backend
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum EpisodeStatusDto {

    BEING_RENDERED = 'BEING_RENDERED',

    RENDERED = 'RENDERED',

    RENDERING_FAILED = 'RENDERING_FAILED'
}